flows/requirements.txt
# Runtime dependencies for the Prisma AIRS + Azure AI Foundry lab Prompt Flow.

# Prisma AIRS Python SDK (aisecurity module)
pan-aisecurity

# Azure ML / AI Foundry client and auth
azure-ai-ml
azure-identity
